refactor: split DatabaseContext into partial classes
